Proposing "Error Message" > "Issue Message"
Per the design principles, it should be impossible for users to cause "errors" in Mattermost. The UI is not reflecting this (https://mattermost.atlassian.net/browse/PLT-3983). 

Propose we make a DISTINCTION between Error Messages (i.e. JavaScript errors and things that are bugs in our system) vs. Issue Messages, which is feedback to the user on how to adjust their input to get the outcome they want. 

We have a UX Defect right now that if you click into an email notification of a DM someone deleted you get this giant word "Error" that makes it seem like the system has failed. 

That giant "Error" word should only be used when the system is actually failing, not when it's operating as intended.