Update config-tls-mattermost.rst
The added note makes it clear that you need not specify TLS settings in both the Mattermost application as well as in the NGINX configuration. If you follow the guide (without these changes) from top to bottom, for say Ubuntu 16.04, then it might be a bit confusing as to why you would be specifying all the TLS settings and certs in the app - if later the guide explains how to do this for NGINX SSL termination.