Add spec.template.metadata.annotations for checksums
This adds the annotations per discussion at #36

In the end, it has them added, but as the `ConfigMap` templates rarely change, they don't actually have an impact. See discussion on #36 as to what and how they are supposed to work, as opposed to how they work with Helm. Also, these may be entirely useless in the future if there is no way to refactor them to be viably as intended.

Closes #36