Unicorn: split disable authorized keys into initContainer
Move the call to `/scripts/disable-auth-keys-write` into an `initContainer` definition.

The pod will not start without this completing successfully. This no longer interferes with Unicorn starting up (which is expensive as it is), which means that the Unicorn services is listening on port 8080 faster, reducing the likelihood of a timeout on the healthcheck causing a pod restart for "no reason".

This is combined with https://gitlab.com/gitlab-org/build/CNG/merge_requests/32

Closes #94