Redis: add persistence capability
Add Redis persistence capability through configuration: `persistence`

Configuration of the disk which data is persisted is done via `disk`. It is currently laid out with only one member, `persistentVolumeClaim`. It is assumed that the administrator will have created the PV/PD prior to the configuration and deployment of the chart, if they chose to use truly persistent storage.

If no `persistentVolumeClaim` is supplied, we'll map in an `emptyDir` volume automatically. This will no persist outside the `Pod` lifecycle, but will persist through container failures or pod restarts. It is limited to `10G` on the node.

```YAML
persistence:
  disk:
    persistentVolumeClaim:
```

Configuration of snapshotting is done via an array of `time` and `change` pairs, which result in lines in `redis.conf` of the format `save ${time} ${changes}`. If the array is empty, then the rendered line will be `save ""` which disables snapshotting entirely.

```YAML
persistence:
  save:
  - time: 60
    changes: 1
``