Docs: re-organize advanced configuration documentation
Reorganize the advanced configuration documentation to a cleaner separation and use.

We want to minimalize both repetition as well as cut and paste actions. This is best served by having advanced configuration provided as YAML snippets that override the same values in `doc/example-config.yaml`, as opposed to inline content in the documentation that needs to be cut and pasted.

To help keep this organized, I've moved `doc/external-db.md` to `doc/advanced/external-db/README.md` and moved the YAML snippet into `doc/advanced/external-db/external-db-snippet.yaml`. Following this model, we should be able to provide clean examples of using any of the chart provided services as an external resource, while keeping the need to update configuration examples to a minimum. (Such has having to updated _every_ `example.yaml` when a property name changes.)

[ci skip]