Docs: update installation documentation to current state of charts
Update the documentation under `doc/installation` to the current state of the charts.

This includes property name changes, and simplification/reduction of items that differ from `doc/example-config.yaml`.

Added `doc/installation/resources.md` for the creation of resources that will be needed (IP / DNS), to be filled in the future.

[ci skip]