Update instructions for configuring the systemd service
Starting from Mattermost 4.8, the platform server can notify systemd
when it is actually started up and ready to accept network connections.

This makes the service status more reliable: systemd will now report the
server as going from "stopped" to "activating" and finally "running".
If the server exits before signaling it is ready, this will be correctly reported as an error to systemd. And executing
`systemctl mattermost start` will return control to the shell when the
server is actually ready–or return with an error exit code in case of
startup failure.