Be consistent in our initContainer naming
We had two patterns in play, making it hard to know which to use. I chose the simpler of the two:

- `${chart}-config`
- `configure`

While looking into this, I also found that we have a mash of `*-dependencies`. The prefix was filled in multiple methods. I have simplified this down to just `dependencies`.