CI: have `cleanup` function include ConfigMaps, remove Secrets
Per review, add `ConfigMap` to the list of items handled in `cleanup()`, and remove `Secret` as it is currently shared by multiple deployments.

Replace `kubectl_cleanup()` entirely, by replacing the function call in `cleanup()` with a condensed selector call through kubectl, in place of an iterative approach over individual types. Care of @ahmadposten