Unicorn: Add support for configuration/use of Omniauth Providers
Incorporate support for configuration of Omniauth within the Unicorn pod.

I've tried to keep this as simple and flexible a change as possible.

Take `gitlab.unicorn.omniauth.providers`, map that into the configuration file, while keeping the secerts in Kubernetes secrets that are populated as a part of `configure` container. The secrets are included in the Deployment via the same logic that handles their population in the ConfigMap.

I have included documentation on the use, with pointers to our public documentation in regards to providers/support/syntax where applicable.

These secrets can not be added via the `shared-secrets` pod, as their sources come from outside the environment.

Closes #335