CI: remove purgeJobs logic from delete()
Remove the `purgeJobs()` logic and call from `delete()` function.

It turns out that when `delete(canary)` is called after `deploy()` in `stable`, the jobs of the `production` deployment are deleted, thus losing logs! It is hard to investigate a problem with migrations when the Job and logs are deleted.

With the addition of the appropriate `cleanup()` logic, `purgeJobs()` is made superfluous, and in this case, proved to be erroneous.