Proposing download links
Propose adding download links to changelog so we can add a new menu option in Desktop App for `Desktop app -> Help -> Check for newer version` which would go the changelog and allow users to notice a new version, and download it right away. 

I've been waiting for 4.1.1 for a long time because of the loading animation. It was very difficult for me to check for a new version. This modification to our changelog process makes it easier to check for newer versions and immediately download once one is found. 

If we do this, propose creating a redirect link from mattermost.com so we don't hardcode to a page, and instead make a link that can be redirected easily in future.