NGINX: replace nginx enabled gating
This replaces `nginx.enabled` with `nginx-ingress.enabled`.

d0d65c7f281087dfda3f06915563acfd56ef208e (nginx-ingress) removed `enabled` flag. This was intended to be replaced with `install` as it was to be an external chart.

This adds back the `enabled` flag to the `nginx-ingress` chart in the `nginx` folder. Sets the default in `charts/nginx/values.yaml`, and `values.yaml` defaults to `true` to enable it.