Sidekiq: add object storage configuration
Add the configuration of object storage to the sidekiq chart.

When Sidekiq does not have the configuration for object storage, there will be instances where you will get error/messages such as `RuntimeErro: object storage is not enabled` when Unicorn/Workhorse _are_ configured with object storage. See #459 for an example where a User's avatar is in object storage, but sidekiq then can't access it.