Unicorn: flesh out OmniAuth settings
Expand the configuration of OmniAuth in the `unicorn` chart, to support all settings.

This alters the implementation of the `providers` list, so that the user must provide the content via k8s Secrets. This is done so that all forms of OmniAuth providers can be used safely, without having to cater to each providers sensitive fields.

Relates to #467