Fix some issues with the script used to bump chart versions
Only fetch gitaly / shell version if the script was not provided a new
gitlab version. (For the case where we are only releasing chart changes)

Also switch the version fetcher to use open-uri for fetching the
versions rather than just Net::Http

This will actually error properly if it gets a bad status code, and will
also follow redirects.

[ci skip]