registry: move `storage` contents to secret
Move `registry.storage` to provide a secret instead of a block to be mapped directly into a ConfigMap, thus providing a safe way to provide sensitive information for external storage.

This adds the following acceptable properties:
```
storage:
  secret:   # name of secret
  key:      # name of key, default `storage`
  extraKey: # name of supplemental key, that will be mounted under the same name.
            #   This is designed for mounting things like GCP credential JSON keys.
```

Additions made to `templates/NOTES.txt` to notify of the deprecation / change in behavior.

Closes #548