Update our changelog to prepare for changelog handling in the release-tools
In order for the release-tools update to work, we need at least one
changelog with the new syntax. So I've gone and left the pre 0.2.0
changes as our weekly ones, and have changed the post 0.2.0 to the
versioned syntax.

I will completely remove the changelog code from this repo after we've
seen a release compeletly work with the new changelog handling.

[ci skip]