redis/minio: set spec.strategy.type for stateful deployments with PVCs
Set the `spec.strategy.type: Recreate` for minio/redis, as they are stateful deployments with PVCs that can not be shared. The introduction of the default to `RollingUpdate` broke upgrades, and required manual user intervention to `scale 0` both deployments in order to upgrade. This ensures the value is present, resulting in explicit behavior.

Adds comment to each chart's `spec.strategy` in regards to setting `spec.strategy.rollingUpdate: null` as a method for frictionless update.