registry: default storage 'delete' to enabled
Even with the changes made in !535 (merged) for #785 (closed),
it is possible that deleting repositories will still fail unless
users ensure that their storage config includes `delete.enabled: true`.

The default when using Minio is `true`. The default within the Omnibus is
`true`. We align these charts to ensure that a value is present, or add
`true` out of the box.

With this change, a user can configure a minimal set of driver constraints,
specifically the driver (gcs/aws/etc). If the user explicitly sets
`delete.enabled`, we then respect that value.

Closes #901