spec: add status code check to sign_in helper
Add a call to check the return code from `click_button 'Sign in'`.

We were performing the submit, and just "assuming" that signin was
successful. This is erroneous because there could be a `500`, or any
other problem at the server. We now check that the status code is `200`
and that the current path is `/` which is the redirect when login is
successful, when `/users/sign_in` is visited directly (per `visit` call).