Doc: update /values.yaml with minimal docs
Update `/values.yaml` with minimal documentation. This primarly adds
documentation links to the various sections, so that users have a quick
reference to our extended documentation.

For the most part, comments in the middle of the values are added as
`## `, so that `# ` can be reserved for commented YAML properties.

Related to #1065