Def of Mattermost staff & Mattermost employee
Drafting explicit definitions of "Mattermost staff", vs. "Mattermost employee" as they're been sprinkled throughout our docs but not yet made explicit. 

There's discussions of "Mattermoster" in different threads and it's not clear if things have gone dead or if people are thinking about follow-ups, in my mind, at the minimum the definitions of Mattermost staff and Mattermost employee are consistent with what we've been using in the past. 

Please flag if anything seems different here than before?  

Other threads

 
- https://gitlab.com/mattermost/proposals/issues/28
- https://github.com/mattermost/docs/pull/2617