Removes deployment replicas key in favor of the HPA
Due to not setting a default for replicas, we fail a `kubeval`
evaluation.  Since we are relying on an HPA anyways, and we set, by
default a minimum and maximum, I don't believe setting replicas here is
necessary at all.  The HPA, after some time, will eventually take
control of the deployment and monitor pod counts.