Set GitLab Monitor versions manually
Fixes an issue where the tag was being set to the gitlab version by the
manage version script.

Monitor isn't updated like our other components, it gets manually
updated in omnibus and here, rather than updated in a file in gitlab-ce
or by release-tools.

Set the version to the current valid tag, and add monitor to our
dependencies.io configuration.