Geo: Add geo-logcursor chart v1
Add a chart under `gitlab/` as `geo-logcursor`.

This chart makes use of `gitlab-geo-logcursor` CNG container from
https://gitlab.com/gitlab-org/build/CNG/merge_requests/220

The chart requires the addition of additional templates to
configured the related tracking database, as well as enable/disable
of related components.

This change _does_ not include database seeding, migrations for
`geo.psql` database, which is currently assumed to pre-exist, and be
up to date on migrations, per the used container.

TODO:
- Documentation
- CI (skipped because #1211 is not closed by CNG!220)

[ci skip]