Disable client-side validation of kubectl apply
When using the command kubectl apply both client and server
validate the resource yaml to a schema. When the client-side
version of kubectl differs too much from the Kubernetes version
on the server (more than one minor version) the Kuberentes API
refuses to accept the resource with the error message:
`Error from server (Forbidden): unknown`

The client-side validation can be skipped with the kubectl flag
`--validate=false`. The Kubernetes API still validates the resource.

This flag is only a work-around until a better way is found to
synchronize the kubectl version used by the gitlab components
and the Kubernetes it is running on (e.g. charts/gitlab!864)

Fixes charts/gitlab#1497