Track the chart version in the gitlab-info
This will allow us to make some pre-hook decisions in the future for
major chart uprades.

Previously we only had done this for Major GitLab changes, but we have
an upcoming major chart release where we may also want to know if users
are upgrading, and this will help us.