Grafana: support for deploying configured grafana
Add ability to deployed upstream Grafana Helm chart
https://github.com/helm/charts/tree/master/stable/grafana

This includes version 3.5.10 of the stable/grafana chart, as a
requirement of the gitlab/gitlab-grafana sub-chart.

The sub-chart provides configuration, dashboards, and an Ingress for
the deployed Grafana instance.

Currently:

- Grafana username is `root`, to be consistent with GitLab. Future
OAuth work may change or disable this.
- Grafana `admin` password is set as a static string in the template
output.
This will need replaced.
- Dashboards are those from Omnibus GitLab. These will need to be made
specifically for the Charts before "complete"

Signed-off-by: Jason Plum <jplum@gitlab.com>
Signed-off-by: Gerard Hickey <ghickey@gitlab.com>