Unicorn: add shared /tmp
Add a shared /tmp mount to the unicorn and workhorse containers.

This addresses #1647, and customer issues.

Changes in GitLab 12.4.0 rails / workhorse (gitlab-org/gitlab!18135)
resulted in the _hard requirement_ of this shared directory.

It is using `emptyDir: {}` to ensure the use of a host-backed working
directory, as `medium: Memory` is _dangerous_ due to the consumption
of the disk counting against `resources.limits.memory`