Unicorn: create tmpdir for Ruby as it is picky
Create a TMPDIR for Ruby / Workhorse, as Rails is picky about the
qualifications of the directory provided by `Dir.tmpdir`

Effectively: must be not world writable, or sticky.

Here, we're making it *both* Sticky and non-world writable (3770)

We'll fix this in the process-wrappers in the CNGs, and this can then be
removed from the chart itself.

Closes #1651