Change recommendation around jobserver and schedulers.
The code changed some time in the past (I can't figure out exactly when)
such that job servers are fully-fledged members of the cluster. This is
necessary and unavoidable, therefore they can also become the cluster
leader. Because of this, they need to have the possibility of running
the schedulers, since without that, there could be no server running the
schedulers if the job server is elected cluster leader, meaning no jobs
get scheduled. Given the only use of cluster leader status in the
codebase now is job scheduling, this change should have no side effects
on any other functionality.