CI: restart task-runner after helm deployment
Restart the task-runner pod of the helm release, after wait_for_deploy.

Our task runner often uses an outdated image in dev, as there is nothing
day-to-day, to cause the pod to be restarted, (even though a new
task-runner latest image is created at least nightly). This isn't a
problem in a production environment, because the images are tagged with
versions, and the version change is enough to cause restarts when needed.

Closes #1593