Ensure that global defaults are set at the top level chart
Global values merge down, not up, so its best to set them at the top
level to ensure they are available everywhere.

For now the only global settings left out are the
enterprise/communityImages. These are located in the nested gitlab
chart, and are currently only need for that chart's subcharts. For now
I am going to leave them there. If we ever want to reference them from a
higher chart, we will have to extract them out as well.