Refactor CRD bootstrapping
The CRD lifecycle management is moved out of the Chart. `crdctl` utility
is introduced to take over this responsibility. CRD bootstrapping with
the Chart is now deprecated and `global.operator.bootstrap` value is no
longer valid.

To use the Operator, the user must use `crdctl` utility first to create
GitLab kind prior installing the Chart. The way that the Operator can be
enabled hasn't been changed.