Update sg_grammar-spelling-mechanics.rst
Issue: 
The technical documentation capitalization style (camel case) differs to the recommendation in the Marketing editorial guide (sentence case). This creates inconsistency in our web properties, can be jarring, and doesn't reflect our unified brand. 

Solution
The technical documentation should use sentence case going forward. With regard to existing content, the proposal is to update all technical documentation in a phased approach per doc set (Admin Guide, User Guide, and so on) to align all the headings. The documentation will be out of sync during this process but it won't be for an extended period of time (roughly two weeks). 

This is linked to: https://github.com/mattermost/mattermost-handbook/pull/78.