Update Performance Tracing to include Plugin API
Recently we aded the ability to trace Plugin API calls and Hooks with prometheus.   We should add an example graph of some possible bottlenecking Hooks or API calls that an admin might want to be aware of when troublehsooting or monitoring their server's performance.  

We wanted to be able to take plugins out of beta, and getting a better sense of performance hits from how plugins interact with the API is an important step forward.