Update Performance Tracing to include Plugin API (#3524)
* Update Performance Tracing to include Plugin API

Recently we aded the ability to trace Plugin API calls and Hooks with prometheus.   We should add an example graph of some possible bottlenecking Hooks or API calls that an admin might want to be aware of when troublehsooting or monitoring their server's performance.  

We wanted to be able to take plugins out of beta, and getting a better sense of performance hits from how plugins interact with the API is an important step forward.

* Update performance-alerting-guide.rst

Added four metrics to Plugin Hooks.

* Update source/administration/performance-alerting-guide.rst

Co-authored-by: Ali Farooq <ali.farooq0@pm.me>

* Update source/administration/performance-alerting-guide.rst

Co-authored-by: Justine Geffen <justinegeffen@users.noreply.github.com>
Co-authored-by: Ali Farooq <ali.farooq0@pm.me>