Redirect filename fix (#5414)
* Redirect filename fix

Replaced code that determined the redirect filename and stripped everything after the first occurrence of `.html`. This code has been replaced with a simpler version that checks whether the entire filename already ends in `.html`; if it doesn't, it appends `.html` to the end of the file. Big shoutout to @neflyte for assisting with this fix.

* redirect corrections