Correct socked based db creation (#5426)
Code labeled "create the database" was actually just a copy/paste of the
"create a user" blob. This is the actual createdb command including
setting the owner of the database to the right user.

Additionally the test code to connect to it does not need (and even
should not have) the username. The example uses peer authentication and
if it is going to work from the mattermost daemon it needs to work
without specifying the user based on the system user ID.