Extra step to see the 'import process' job status (#5457)
The "Migrating from Slack" section is improved with one new step to view the status `mmctl import process` job. This extra step allows the user to determine success/error of the slack migration import before returning to the webapp. If there was a error an error message, it is shown to help with debugging.

Fixes #5449