Small fixes identified while following (#5832)
* Small fixes identified while following

I tried to follow the instructions today and recognized a few issues with this documentation.

If you do not move the "bulk-export-attachments" folder to a "data" subfolder, the import process is not able to find the attachments.
Also the filename "mattermost-bulk-export.zip" was wrong (should be import).

I also saw that this document uses content backlinks top, others don't. What's the default setting of that option? I mean the documentation looks a bit different anways depending on what site you are (anchor links on the left, anchor links on the right, clickable headings yes/no)...

Changing the "backlinks" option to "none" here fixes this problem and it's much more intuitive IMHO if you want to share the link to a section, but then again, you cannot go back to the TOC with just one click.

* Update source/onboard/migrating-to-mattermost.rst

Co-authored-by: Justine Geffen <justinegeffen@users.noreply.github.com>
Co-authored-by: Carrie Warner (Mattermost) <74422101+cwarnermm@users.noreply.github.com>