Support request: Make section headings sharable (#6188)
* Support request: Make section headings sharable

The sections that detail how to migrate from Slack using the mmetl tool and bulk import included bolded headings that are challenging for the Support team to share with customers. Converted these headings so that they render as proper anchors that Support can share publicly. Adjusted heading formatting across the page to compensate.

* Heading adjustments for consistency

* Update migrating-to-mattermost.rst

---------

Co-authored-by: Justine Geffen <justinegeffen@users.noreply.github.com>