Fix wrong filename and format
This command does not work, the processed import file always gets an ID generated and prepended to the file and needs to be specified for the import process command. Additionally, JSONL Files cannot be uploaded directly, only ZIP files are supported.

The preview didn't work for me here, so I'm not 100% sure I got it right as far as formatting is concerned, sorry.