Clarify use of read replicas
The docs suggest that reads are evenly distributed between the master DB and the read replica DBs. Based on both the server code and the observed behaviour, I do not believe that is the case - I believe that (if any read replicas are available) read queries are distributed among the read replicas only. This is a useful distinction to keep in mind for site admins.