[feat] Migrating to GitHub Actions
Introducing cd.yml and ci_master.yml worfklows, to tackle safe deployment to AWS S3 service.

```
The new trigger workflow_run was introduced to enable scenarios that require building the untrusted code and also need write permissions to update the PR with e.g. code coverage results or other test results. To do this in a secure manner, the untrusted code must be handled via the pull_request trigger so that it is isolated in an unprivileged environment.
```