Patch 1 (#7017)
* listen on IPv6 as well as IPv4

Added a line to also listen on IPv6. Without this line, I was unable to connect to my freshly installed mattermost server, because IPv6 seemed to be automatically chosen, but nginx didn't listen on it.

* removed ssl directive

the ssl directive is deprecated and "listen ... ssl" should now be used, according to nginx. removing the statement "ssl on;" doesn't affect behaviour, because ssl is already enabled by the listen directive.