Update metrics docs (#7356)
I tried to automate generating a list with all metrics registered, but
that ended up being quite cumbersome, so this is the result of a
semi-automated generation plus some manual post-processing.

My idea was to use the Gather function in the registry:
https://pkg.go.dev/github.com/prometheus/client_golang/prometheus#Gatherer
and write a script that simply called that. However, it's not as easy:
- The `go_sql_*` metrics need to be registered explicitly. This is
  doable, it's just those ones.
- The `*Vec` metrics need to be observed at least once so that those
  metrics are reported back by `Gather`. This is error-prone, there are
  multiple ones, and doing this manually it's terrible

You can see the test I wrote in this commit:
https://github.com/mattermost/mattermost/commit/6238fce4e3835974e99c00d7f11e0ef6dc80586a
That prints the metrics in the same order we had in the docs, so that it
was easy to compare if my test was missing some registered
metrics (i.e., if there were some *Vec metrics whose
Observe/Increment/Decrement functions I had not called). That happened
quite a lot, so it was a very manual process, and we'd need to maintain
that list of calls to the Observe/Increment/Decrement functions. Not
good.

If we want to really automate this, I see two alternatives:
1. Use custom functions wrapping the `prometheus.New*` and
   `MustRegister`, so that we can collect the names and descriptions in
   a list we can later export.
2. Use `reflect` to inspect the calls to `MustRegister` and collect the
   names and descriptions of the passed arguments.