Improve Copilot + local AI setup instructions (#7413)
* Improve Copilot + local AI setup instructions

* Update source/configure/enable-copilot.rst

* 📝 (academy-copilot-setup.rst): Add new file to include Mattermost Academy badge
📝 (enable-copilot.rst): Include Mattermost Academy badge and update API URL instructions
The new file academy-copilot-setup.rst was added to include a Mattermost Academy badge that links to a tutorial on setting up and configuring Mattermost Copilot with multiple LLMs. This badge was then included in enable-copilot.rst to provide users with a direct link to the tutorial. The instructions for entering the API URL were also updated to clarify the URL format for services other than Ollama.

---------

Co-authored-by: Carrie Warner (Mattermost) <74422101+cwarnermm@users.noreply.github.com>