MM-63660: Added client-side filtering capability for upgrade notes and changelog (#7848)
We add JS and CSS code to dynamically filter the rendered
HTML files in the upgrade notes and changelog pages.

I tried several approaches here before setting down on this one.

My initial approach was to convert the .rst file to a JSON file
during the build process and then dynamically load the JSON file.
But that had issues because the .rst contents had links and code
which needed to be formatted correctly.

The second approach was to see if it's possible to template
this on the server side. This would have been the right solution
here because the current solution requires parsing through HTML
which is brittle. But unfortunately, this would require changing
the architecture of the app from a static HTML app to being served
with Python. While that is possible, it is out of scope.

Therefore, we settle down on this approach where we include JS
to filter through the rendered HTML files and apply a bit of CSS
to polish the UI.

The result works fine. But it's likely a bit brittle because of
the lack of ability to control the rendered HTML.

Most of it assisted by Claude

https://mattermost.atlassian.net/browse/MM-63660

Co-authored-by: Carrie Warner (Mattermost) <74422101+cwarnermm@users.noreply.github.com>